requirements.txt
pyttsx3>=2.90
pydub>=0.25.1
paho-mqtt>=1.6.1
paramiko>=2.11.0
redis>=4.3.4
aiohttp>=3.8.1
crypto>=1.4.1
pycryptodome>=3.15.0
TwitterAPI>=2.7.13
mysql-connector-python==8.0.29
requests>=2.28.1
urllib3>=1.26.11
deepgram-sdk>=0.3.0
twilio>=7.12.0
scp>=0.14.4